requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Additional dependencies
typing-extensions==4.8.0
python-dotenv==1.0.0
